# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema Benefícios",
        "description": "Documentação da API do Sistema de Benefícios",
        "contact": {
            "name": "Luiz Chimenes",
            "email": "gustavo.chimenesp@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza o login do usuário",
                "description": "Autentica um usuário e retorna um token JWT",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "senha"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@empresa.com"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Login realizado com sucesso"
                                        },
                                        "dados": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "tipo_token": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                },
                                                "expira_em": {
                                                    "type": "integer",
                                                    "example": 3600
                                                },
                                                "usuario": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "nome": {
                                                            "type": "string",
                                                            "example": "João Silva"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "joao@empresa.com"
                                                        },
                                                        "tipo": {
                                                            "type": "string",
                                                            "example": "admin"
                                                        },
                                                        "departamento": {
                                                            "type": "string",
                                                            "example": "TI"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Credenciais inválidas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza logout do usuário",
                "description": "Invalida o token JWT atual",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Renova o token JWT",
                "description": "Retorna um novo token JWT válido",
                "operationId": "2c27d311a2876560fcb5a754b556cdc5",
                "responses": {
                    "200": {
                        "description": "Token renovado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "tipo_token": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                },
                                                "expira_em": {
                                                    "type": "integer",
                                                    "example": 3600
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/perfil": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Retorna informações do usuário logado",
                "description": "Retorna os dados do perfil do usuário autenticado",
                "operationId": "f812ecf7f6db8245b4a29a8f216c0c16",
                "responses": {
                    "200": {
                        "description": "Dados do usuário retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nome": {
                                                    "type": "string",
                                                    "example": "João Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao@empresa.com"
                                                },
                                                "tipo": {
                                                    "type": "string",
                                                    "example": "admin"
                                                },
                                                "departamento": {
                                                    "type": "string",
                                                    "example": "TI"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}